fix: Resolve all compilation warnings and clippy issues
- Fixed unused imports across all packages
- Applied cargo fix recommendations
- Formatted inline variables in format strings
- Cleaned up unused variables and parameters
- Ensured all code passes cargo check without warnings

+++ hanzo-libs/hanzo-sqlite/src/mcp_server_manager.rs
@@ -137,7 +137,7 @@ impl SqliteManager {
             env.map(|e| serde_json::to_string(&e).unwrap())
                 .unwrap_or_else(|| "{}".to_string()),
             if is_enabled { 1.to_string() } else { 0.to_string() },
-            (id as i64).to_string(),
+            id.to_string(),
         ])?;
         match rows.next()? {
             Some(row) => Ok(MCPServer {
